AI Module: Response Parser (#112)
The PR ports over the ResponseParser.ts class from the JS SDK.

Details
- Implemented virtually the same public interface. Same public methods,
return type might be slightly different.
- Response Parser can parse an adaptive card or plan JSON string into a
`AdaptiveCard` or `Plan` object respectively. These two functionalities
are thoroughly tested.
- If no plan JSON is found in the response, then it defaults to parsing
logic verbatim to JS SDK. Will need to clarify plan structure in the
response, and will take time thoroughly test it. Because this is not
integral to have the existing samples working, I will leave that for a
later in the interest of time.

Response Parser TODOs:
https://github.com/microsoft/botbuilder-m365/issues/113