Turn State Infrastructure  (#232)
Added support for the turn state. resolves #209 

Additions
- Turn State
- Turn State Manager
- Conversation history
- Integration of Turn State with AI, Application, OpenAIPlanner classes

Missing that will come in subsequent PRs:
- `Application` unit tests
- `AI` class unit tests

A few major differences between this and the JS SDK:
- Dropped `Default` from `DefaultTurnState` to get `TurnState`. Not sure
why `Default` was there in the first place, I presume it's to signal to
the user that this component is configurable. We do that implictly in C#
with interfaces (ex. `ITurnStateManager`) and unsealed class (ex.
`TurnState`).
- `Application` class now has two generic parameters, `TState` and
`TTurnStateManager`. This was to solve the problem that there is
virtually no covariance in C# generic types (i.e
`ITurnStateManager<TurnState> var = TurnStateManager<ChildTurnState>`
throws compiler errors, even though `ChildTurnState` extends
`TurnState`).