[C#] feat: seal/internalize non-essential classes (#537)
## Linked issues

closes: #531  (issue number)

## Details

Provide a list of your changes here. If you are fixing a bug, please
provide steps to reproduce the bug.

Internalized
- OpenAIClient: We don't expect end users to use our open ai client to
directly call the api. They should use other existing libraries for
this.
- OpenAIClientOptions: Part of the OpenAI client
- DoCommandActionData: a helper class

Sealed
- A bunch of classes we don't expect the end user to extend.


The motivation behind aggressively sealing/internalizing classes is to
prevent end users from relying on components we don't explicitly expect
them using. Eg. The OpenAIClient's sole purpose is to service the
OpenAIModerator, even though end users can use it to directly call the
API.

If there's a legitimate reason or customer ask to unseal/publicize a
class then we can easily do that without break anything. Doing the
opposite is a big no-no as it is a breaking change.

#### Change details

> Describe your changes, with screenshots and code snippets as
appropriate

**code snippets**:

**screenshots**:

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (we use
[TypeDoc](https://typedoc.org/) to document our code)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below