[JS] feat: add teams sso bot sample (#838)
## Linked issues

closes: #836 

## Details

Add Teams bot SSO prompt based bot auth sample. The sample is ready for
local debug and remote deploy.
When user sends anything to the bot for the first time, it will single
sign-on current user and send the length of acquired token to the user.

Note:
Like other new samples, since the latest code is not published, people
need to update package.json manually to reference a local package before
try this sample.

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (we use
[TypeDoc](https://typedoc.org/) to document our code)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes