[JS] feat: Who bot sample, add user sign in method for end user (#911)
#### work in progress
- [x] Update the sample to include more graph calls and test it

## Linked issues

closes: #890 #895

#### Change details

- Added Who bot sample
- Added `getTokenOrSignInUser` method in `Application` class. It is
meant to be the single catch-all method to sign a user in and get the
token. If the user is signed in it simply returns the token. If not, it
triggers the sign in flow. Starting the sign in flow will only work if
the incoming activity supports it (ex. `message`,
`messageExtension/fetch`...etc). This method will work in action
handlers as well.

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (we use
[TypeDoc](https://typedoc.org/) to document our code)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes