[PY] feat: add prompt manager (#1233)
## Linked issues

closes: #1065

## Details

1. Add `PromptManager` and corresponding classes
2. Accessing `DataSource.name` will throw error with `TextDataSource`
instance. Update `TextDataSource` and `DataSource` class to expose the
`name` property to fix this issue.
3. Add `from_dict` method to some dataclasses to allow them initialize
from a json file

Note: augmentations and actions related logic are not included in this
PR as they're not ready yet. I have added comments to related issues as
follow up items.

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

---------

Co-authored-by: Alex Acebo <aacebowork@gmail.com>