[JS] fix: ME Auth card title is a fixed string (#1444)
## Linked issues

closes: #1445  (issue number)

## Details

Provide a list of your changes here. If you are fixing a bug, please
provide steps to reproduce the bug.

#### Change details
* Updating `MessageExtensionAuthenticationBase` class to take in "title"
and "text" paramters in it's constructor which it then uses to populate
the auth invoke response. If not provided, it uses default hardcoded
strings.
* I've only updated this for `OAuthMessageExtensionAuthentication` as
the `TeamsSsoMessageExtensionAuthentication` settings (TeamsSsoSetting)
does not currently support `title` and `text` properties.

**screenshots**:

I tested it in the ME OAuth bot and the title updated correctly. The
text however did not, it is most likely a frontend issue. I've reached
out to them about it.

![image](https://github.com/microsoft/teams-ai/assets/115390646/aafa8753-80bb-4348-a95d-a87d9d29fc12)


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below