[PY] feat: Twenty Questions (#1577)
## Linked issues

closes: #316

## Details

A python implementation of the existing JS "Twenty Questions" sample
bot.
The bot leads the user through the twenty questions guessing game, with
the user attempting to guess a secret word via successive guesses.
Driven through the Teams toolkit, the sample launches a browser window
where a instance of teams chat runs the bot locally. See screenshot for
interaction.

#### Change details

Ported the code found in
`/js/samples/03.ai-concepts/a.twentyQuestions/...` to Python, with usage
of the existing Python implementation of things like`Application`,
`ActionPlanner`, `context`, and so on.

**code snippets**:
None.

**screenshots**:

![image](https://github.com/microsoft/teams-ai/assets/50149972/f0ebffe0-e42c-40ea-a5f3-c056a9c2ddad)

## Attestation Checklist

- [X] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------

Co-authored-by: Lily Du <lilydu@microsoft.com>