[JS] feat: Update Citations behavior in SayCommand (#1591)
## Linked issues

closes: #1587 

## Details

- (Temporarily) remove `citations.url` from being added to the Citations
object sent to Teams. We are doing this because right now, the link used
is not accessible (will result in a permissions denied view)
- To work around this, developers can extend SayCommand to apply their
own url manually. Directions are already discussed in getting started
docs.
- Update `formatCitationsResponse` to handle references up to three
digits in length (e.g. [doc123])
- Add new utility, `getUsedCitations`, which will reference the text
(e.g. 'testing [1], something [2]') and compare to the citations array.
Any citations that are not referenced in the text (e.g., [3]), will be
removed from the citations array.
- Cleanup SayCommand entity code and how objects are added into the
activity


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------

Co-authored-by: Corina Gum <>