Copy activity to new activity when continuing conversation
When `Application` is instantiated with `long_running_messages=True`, the incoming request is converted to a proactive conversation, which creates a new activity. The original code has a bug where the original activity's attributes aren't copied to the new activity, which results in the original activity intent not being handled. Fix this by copying the activity in a new callback function which wraps the original callback function.