[PY] fix: Do not throw error if there is no text content (but attachments) (#1850)
## Linked issues

closes: #1777

## Details

If there is a message that has no text content, but it does have some
non-textual attachments, we currently throw. However, the message could
have attachments that could be used for further processing by the
pipeline. As such, it doesn't make sense to throw an error in this case.

#### Change details

Before the error:

![image](https://github.com/user-attachments/assets/f61b38f4-1c71-43f2-8d33-f37af13407db)

After the error:


https://github.com/user-attachments/assets/01a1f260-8bde-4d01-bd55-a660c1a905fa

Note: 
I didn't see any existing tests that tested this particular path (esp
using `options.ai`). Any pointers on building that out? I started adding
to `test_app.py` but wanted to know if there were easy ways of mocking
out the AI layer out. Hence no tests yet, but happy to add it if it's
not that hard.

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------

Co-authored-by: Lily Du <lilydu@microsoft.com>