[PY] feat: SSO-token-exchange (#1867)
## Linked issues

closes: #1870 

## Details

Focuses on token concepts as relates to SSO token exchange, as started
in the py-sso branch started by @lilyydu
https://github.com/microsoft/teams-ai/tree/lilyydu/py-sso. Adds basic
exchange concepts to bring the py-sso branch somewhat in line with other
teams-ai implementations on this topic. Also starts a few tests on the
topic for later enhancement.

#### Change details

**code snippets**:

**screenshots**:

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below