Updated StreamingResponse class to support queue based delivery of chunks to client
- Added an outbound activity queue to the StreamingResponse class.
- Made protocol fixes identified in latest PR submission.
- Updated LLMClient class to use new queue based chunking methods.

These changes avoid issues with the non-async events fired from the model. Chunks can come in as slow or as fast as they want from the model and they will be delivered as quickly as they can to the client. Chunks may be combined prior to deliver to the client.