[JS] feat: ChatCompletion - Tools (#1916)
## Linked issues

closes: #1576 

## Details
Implement Tools Augmentation which will allow the client to call tools
(actions) for OpenAI-like models. Since not all models use the term
'tools' for calling functions, we will continue to use `actions` as a
general term. This work does not include adding ToolsAugmentation to
other models.

#### Change details

- Implement JS `ToolsAugmentation` - enable this augmentation in
`config.json` to facilitate the AI model indicating when to make tool
(action) calls
- Bump `openai` to 4.41.1 due to an import error
- Store `actionId` (the equivalent to `tool_call_id` for OpenAI tool
calls to `DO` commands
- `OpenAIModel.ts` will convert `actions.json` to tools and send them to
the model, which when receiving `tool_calls` from the model will store
those as action calls in conversation history in `LLMClient.ts`.
- `ActionPlanner.ts` via `AI.ts` `run` command will call the tools as DO
commands and save the output / send back to model in `OpenAIModel.ts`

#### Other:
- Update TeamsAI diagram and add it to `CONCEPTS.md`
- Unit tests coverage increase 


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

---------

Co-authored-by: Corina Gum <>