[PY] feat: SSO - auth-sign-in (#1976)
## Linked issues

closes: #1318

## Details
Added the get_token_or_sign_in_user() method. It is the interaction
entrypoint for the sso authentication.

It is modeled after the JS version of this functionality. 

If the user is signed in, get the access token.
If not, triggers the sign in flow for the provided authentication
setting name and returns.
In this case, the bot should end the turn until the sign in flow is
completed.

#### Change details

> Describe your changes, with screenshots and code snippets as
appropriate

**code snippets**:

**screenshots**:

## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------

Co-authored-by: Lily Du <lilydu@microsoft.com>