[repo] docs: Update README.md fixed a typo (#2114)
#minor Corrected "are" to "is" in [README.md].

This pull request addresses a minor typo found in the repository. The
typo has been corrected to improve clarity and maintain the quality of
the documentation.

This change is purely cosmetic and does not affect functionality.

## Linked issues

closes: # (issue number)

## Details

Provide a list of your changes here. If you are fixing a bug, please
provide steps to reproduce the bug.

#### Change details

> Describe your changes, with screenshots and code snippets as
appropriate

**code snippets**:

**screenshots**:

## Attestation Checklist

- [ ] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below

---------

Co-authored-by: Corina <14900841+corinagum@users.noreply.github.com>