[C#] fix: When streaming operation failed the exception is suppressed. (#2122)
## Linked issues

closes: #minor

## Details
Exceptions thrown in `DrainQueue()` are being supressed as the method
runs on a worker thread and the corresponding `Task` is never resolved
by the main thread.

With the fix, when `this._queueSync` is awaited in `EndStream()`, an
exception will be thrown if the corresponding `Task` is faulted (i.e. an
exception is thrown from `DrainQueue()`.

#### Change details
* Removed code that sets `this._queueSync` to null when queue is empty.
This means that `this._queueSync` can be either null or assigned a Task
at any given time.
* If `this._queueSync` is not null then an additional check
`this._queueSync.isTaskCompleted` is made to check whether `DrainQueue`
Task has completed or not.

If streaming operation fails, the exception will be thrown:


![image](https://github.com/user-attachments/assets/403b6003-7df4-4684-b433-dd507b91cb4c)


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes