[C#] feat: Streaming - Citations & Sensitivity Label (#2147)
## Linked issues

closes: #1970

## Details

This will work slightly differently than the previous non-streaming Plan
flow. Citations and its respective sensitivity labels are added per each
text chunk queued. However, these will only be rendered in the final
message (when the full message has been received).

Rather than exposing the SensitivityUsageInfo object as an override on
the PredictedSayCommand, the label can now be directly set as usageInfo
in the AIEntity object along with the AIGenerated label and the
citations.

**screenshots**:

![image](https://github.com/user-attachments/assets/81c0f3dd-f958-4b4b-8cbf-56bf36f24f78)


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

---------

Co-authored-by: lilydu <lilydu+odspmdb@microsoft.com>