[PY] feat: Streaming + Sample (#2141)
## Linked issues

closes: #1967, #1970

## Details

- Added streaming support for Python with an associated sample. As
ChefBot does not exist in Python, streaming is demonstrated through the
ListBot instead.
- Added Powered by AI features for the final chunk (feedback loop,
citations, sensitivity label, gen by AI label)
- Designed a custom `PromptCompletionModelEmitter` class to manage
streaming events and handlers
- **Feedback Loop:** The flag is passed in slightly differently from JS
and C#. This is because the AI class was never being passed in as a
method param for `Planner.continue_task `. Adding this change now as an
optional param will lead to compiling errors for all developers with
custom planners (parent class needs to be updated). This also leads to a
couple of cyclical dependencies. Instead of internally piping this flag,
it is now exposed as an _option_ in the `ActionPlanner`. This is (a)
more obvious to configure, (b) will not introduce a sudden breaking
change to non-streaming developers, and (c) provides more control for
the developer. Although this flag can deviate from the one set in the
`AI` class, it is not possible for both flows to run at once, so there
won't be any conflicts.
- **Citations and Sensitivity Label**: This will work slightly
differently than the previous non-streaming Plan flow. Citations and its
respective sensitivity labels are added per each text chunk queued.
However, these will only be rendered in the final message (when the full
message has been received). Rather than exposing the
`SensitivityUsageInfo` object as an override on the
`PredictedSayCommand`, the label can now be directly set as `usageInfo`
in the `AIEntity` object along with the AIGenerated label and the
citations.

Additional items for parity:
- Added temporary 1.5 second buffer to adhere to 1RPS BE service
requirement. Consequently, if the message is small, the message will not
look like it is being streamed. Recommended command to visualize the
feature is "tell me a story and render this as a large text chunk".
- Added reject/catch handling for errors
- Added entities metadata to match GA requirements. This will log a few
warnings from the Botbuilder side, until it is added on their side, in a
few months.

**screenshots**:


![image](https://github.com/user-attachments/assets/4e6c1606-dfa0-49f0-a652-66174eb7c9be)


## Attestation Checklist

- [x] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

---------

Co-authored-by: lilydu <lilydu+odspmdb@microsoft.com>