[JS] feat: streaming support for Tools Augmentation (#2195)
## Linked issues

closes: #2197

## Details

This update adds support for properly calling tools with streaming
support. It doesn't solve for the other augmentation types as that's
much more challenging.

#### Change details

The challenge with streaming + tool use is that we need to make multiple
calls to the model and we can only have a single stream open to the
teams client. To address this we end up leveraging the fact that the
Streamer is being cached to temp state and we let the tool call happen
as it normally world. To the tool call its as if its a non-streaming
message but the Streamer object is kept in memory (temp state) waiting
for the tool call to complete. We a new request is made to the
ActionPlanner, the created LLMClient will see that it's already created
a Streamer and attach to that instance.

There was a bit of book keeping that needed to be done but it's a
straightforward change.

**code snippets**:

**screenshots**:

## Attestation Checklist

- [x ] My code follows the style guidelines of this project

- I have checked for/fixed spelling, linting, and other errors
- I have commented my code for clarity
- I have made corresponding changes to the documentation (updating the
doc strings in the code is sufficient)
- My changes generate no new warnings
- I have added tests that validates my changes, and provides sufficient
test coverage. I have tested with:
  - Local testing
  - E2E testing in Teams
- New and existing unit tests pass locally with my changes

### Additional information

> Feel free to add other relevant information below