Fix tool call outputs being overwritten
Add a sublevel to the list of tools called as part of a run (this list is stored on assistants_planner.py `tool_map:dict`), so that their result values are not overwritten by subsequent calls to the same tool within the same run.

The current implementation did not account for multiple calls to the same tool as it did not keep track of the call id (`call.id`) which is different for each call, regardless of which tool. The list was only being indexed by `tool_id`. The result was that subsequent calls to the same tool would overwrite the result of the previous call.

The fix consists on adding a second level to the dictionary of the tool calls (which contains the return values), so instead of being (pseudocode) `tools[tool_id]` it becomes `tools[tool_id][call_id]`. This way the return value of each and every call is retained.