Remove numbering in filenames (#2425)
Removes numbering from filenames in docs as discussed. This is so that
our links don't have numbering in them, and so that we don't have to
manage ordering of file-names

Also fixes a bunch of broken links. While doing this, I ended up
creating a script that checked for validity of links. I'm adding that as
a github workflow actino.